refactor: rename service file
